### 2018-08-05 #872 @whedon
[PRE REVIEW]: medical image visualization library in python
**Submitting author:** @raamana (<a href="http://orcid.org/0000-0003-4662-0558">Pradeep Reddy Raamana</a>)
**Repository:** <a href="https://github.com/raamana/mrivis" target ="_blank">https://github.com/raamana/mrivis</a>
**Version:** 0.3
