### 2018-08-05 80dba8b
add funding acknowledgement
