### 2018-08-05 6bcc5ba
Update paper.md
