### 2018-08-11 008a571
using a customizable value to represent background
