### 2018-08-12 #872 @arokem
I have a question regarding the title of the article: the software is called `mrivis` and the title on the repo refers only to neuroimaging. It also seems that the file formats supported by the software are neuroimaging-specific. I wonder whether the title of the article could be made more specific to reflect this?
