### 2018-08-13 #872 @raamana
Thanks Ariel for agreeing to edit and starting the review. Appreciate it.

Yes, the docs need to improved, and I plan to fix it shortly.

Just to clarify, yes It’s mostly tested for neuroimaging data so far, but there is absolutely nothing that’s not generic to other medical image data i.e. any type of 3D/4D data that can be read by nibabel or similar library.
