### 2018-08-13 #872 @raamana
Also, I've just updated the paper.md, is it possible to regenerate the PDF for review? cc @arfon
