### 2018-08-13 a1cc248
fixing CI tkinter no display error
