### 2018-08-13 cd4c131
removing incorrect kwarg, and comments in tests
