### 2018-08-14 1234b81
improved error msgs
