### 2018-08-15 5cb2958
generalizing read_image to check for 4D image shape
