### 2018-08-15 9288a5f
better validation of fixed dim
