### 2018-08-15 144ca43
reducing confusion with an option for a single mask: roi labels, which could be doubled as a simple mask if need be
