### 2018-08-15 44108fb
tightening the handling of two masks
