### 2018-08-15 3a1039b
generalizing the way summary clusters are computed
