### 2018-08-15 9a03ae1
use of member variable
