### 2018-08-15 7b4b79e
generic way to handle unrolling
