### 2018-08-15 0652fab
adapting read_image for 4D images
