### 2018-08-15 38fdb25
untested pseudocode impl. for clustering per ROI list
