### 2018-08-15 484d302
basic implementation of exporting to gif from slice picker
