### 2018-08-15 533a606
improving implementation and suppressing warnings
