### 2018-08-15 4bd902c
test to ensure saving to gif works
