### 2018-08-15 64ebf44
improving docs
