### 2018-08-16 dbd6b42
much improved docs
