+++ docs/usage.rst
@@ -1,7 +1,7 @@
-=====
+------
 Usage
-=====
+------
 
-To use mrivis in a project::
-
-    import mrivis
+For usage instructions on how to use `mrivis`
+ - to check alignment, see :doc:`check_alignment`
+ - to use different classes, see :doc:`classes`
