+++ docs/classes.rst
@@ -1,0 +1,16 @@
+---------
+Classes
+---------
+
+A detailed usage illustration is available in the following `demo notebook <https://nbviewer.jupyter.org/github/raamana/mrivis/blob/master/docs/example_notebooks/mrivis_demo_Collage_SlicePicker_classes.ipynb>`_
+
+
+API Reference for Classes
+----------------------------
+
+.. automodule:: mrivis
+   :members: SlicePicker, Collage, Carpet
+   :undoc-members:
+   :show-inheritance:
+
+
