### 2018-08-16 e5adc73
better handling of exception
