### 2018-08-17 ed2392e
tightening the behaviour - not using self inside
