### 2018-08-17 a3c23d4
tightening the behaviour with static method - not using self inside
