### 2018-08-17 b04d46c
cleaning - removing unused variables
