+++ mrivis/tests/test_mrivis.py
@@ -4,8 +4,7 @@ import matplotlib
 # to avoid use of Xwindows backend on CI servers
 matplotlib.use('Agg')
 
-from matplotlib import pyplot as plt
-from os.path import join as pjoin, abspath, realpath, basename, dirname, exists as pexists
+from os.path import join as pjoin, realpath, basename, dirname, exists as pexists
 from mrivis import checkerboard, color_mix, voxelwise_diff
 from mrivis.utils import scale_0to1, read_image
 import numpy as np
