+++ docs/classes.rst
@@ -1,6 +1,6 @@
----------
-Classes
----------
+---------------------------
+Visualization Classes
+---------------------------
 
 A detailed usage illustration is available in the following `demo notebook <https://nbviewer.jupyter.org/github/raamana/mrivis/blob/master/docs/example_notebooks/mrivis_demo_Collage_SlicePicker_classes.ipynb>`_
 
