+++ _sources/index.rst.txt
@@ -1,0 +1,23 @@
+Welcome to mrivis's documentation!
+======================================
+
+Contents:
+
+.. toctree::
+   :maxdepth: 2
+
+   readme
+   installation
+   usage
+   check_alignment
+   classes
+   API
+   contributing
+   authors
+
+Indices and tables
+==================
+
+* :ref:`genindex`
+* :ref:`modindex`
+* :ref:`search`
