+++ _sources/usage.rst.txt
@@ -1,0 +1,7 @@
+------
+Usage
+------
+
+For usage instructions on how to use `mrivis`
+ - to check alignment, see :doc:`check_alignment`
+ - to use different classes, see :doc:`classes`
