+++ _sources/readme.rst.txt
@@ -1,0 +1,20 @@
+------
+mrivis
+------
+
+.. image:: https://img.shields.io/pypi/v/mrivis.svg
+.. image:: https://api.codacy.com/project/badge/Grade/370b1b78383b40b99eb15ae0490cfbdb
+    :target: https://www.codacy.com/app/raamana/mrivis?utm_source=github.com&amp;utm_medium=referral&amp;utm_content=raamana/mrivis&amp;utm_campaign=Badge_Grade
+
+`mrivis` offers:
+
+ - tools for comparison of spatial alignment of two 3D MRI scans (T1, T2, PET etc) using checkerboard plots and other comparison methods: :doc:`check_alignment`
+ - classes to develop customizable visualizations of 3D and 4D images - `demo <https://nbviewer.jupyter.org/github/raamana/mrivis/blob/master/docs/example_notebooks/mrivis_demo_Collage_SlicePicker_classes.ipynb>`_
+
+
+
+
+.. image:: flyer_option_matrix.png
+
+
+
