+++ docs/API.rst
@@ -1,10 +1,0 @@
-
-API Reference
---------------
-
-A tutorial-like presentation is available at :doc:`usage`, using the following API:
-
-.. automodule:: mrivis
-   :members: checkerboard, voxelwise_diff, color_mix
-   :undoc-members:
-   :show-inheritance:
