+++ docs/alignment_API.rst
@@ -1,0 +1,11 @@
+
+----------------------
+Alignment Check - API
+----------------------
+
+A tutorial-like presentation is available at :doc:`alignment_usage`, using the following API:
+
+.. automodule:: mrivis
+   :members: checkerboard, voxelwise_diff, color_mix
+   :undoc-members:
+   :show-inheritance:
