+++ docs/conf.py
@@ -56,6 +56,8 @@ templates_path = ['_templates']
 # The suffix of source filenames.
 source_suffix = '.rst'
 
+autoclass_content = 'both'
+
 # The encoding of source files.
 #source_encoding = 'utf-8-sig'
 
