### 2018-08-17 e80d507
updated docs
