+++ _sources/usage.rst.txt
@@ -3,5 +3,5 @@ Usage
 ------
 
 For usage instructions on how to use `mrivis`
- - to check alignment, see :doc:`check_alignment`
+ - to check alignment, see :doc:`alignment_usage`
  - to use different classes, see :doc:`classes`
