+++ _sources/alignment_API.rst.txt
@@ -1,0 +1,11 @@
+
+----------------------
+Alignment check - API
+----------------------
+
+A tutorial-like presentation is available at :doc:`alignment_usage`, using the following API:
+
+.. automodule:: mrivis
+   :members: checkerboard, voxelwise_diff, color_mix
+   :undoc-members:
+   :show-inheritance:
