### 2018-08-17 24500ff
updating readme with links docs site
