### 2018-08-17 9cc4adc
improved readme
