### 2018-08-17 1fb11f7
fixing docs-related readme
