### 2018-08-17 #897 @raamana
Hi @miykael, the docs website for you to get a sense of what it is. I am finalizing the implementation for one another class Carpet, but what it is now should give you a very good sense of what's happening.
