### 2018-08-19 19da8f2
moving save_figure, and removing unnecessary path modifications
