### 2018-08-19 069552a
code simplification and cleanup
