### 2018-08-19 7c9feaf
bug fix: transpose=False was never tested, would not have returned any data
