### 2018-08-19 a51e20a
better/simplified class design
