### 2018-08-19 55b19d1
improved parameter validation and state initialization
