### 2018-08-19 5209c4e
new method to validate input image
