### 2018-08-19 6586703
improved checks on the dimension to be fixed
