### 2018-08-19 9af4cc5
docs for show method
