### 2018-08-19 a0674e1
making the return values testable
