### 2018-08-19 7a38918
single show method
