### 2018-08-19 91326e3
high-level method to cluster rows in carpet - elegant implementation
