### 2018-08-19 15e83c7
actual implementations for clustering/summarizing
