### 2018-08-19 7ffa194
correct roi_mask handling
