### 2018-08-19 7b21562
code style improvement
