### 2018-08-19 cff9322
method to save viz to disk
