### 2018-08-19 46948d3
convenience repr method
