+++ mrivis/__init__.py
@@ -11,7 +11,7 @@ __all__ = ['checkerboard', 'color_mix', 'voxelwise_diff', 'collage', 'aseg_on_mr
 from sys import version_info
 if version_info.major > 2:
     from mrivis.workflow import checkerboard, color_mix, voxelwise_diff, collage, aseg_on_mri
-    from mrivis.base import Collage, SlicePicker, MiddleSlicePicker, MidCollage
+    from mrivis.base import Collage, SlicePicker, MiddleSlicePicker, MidCollage, Carpet
     from mrivis import color_maps
 else:
     # from .mrivis import checkerboard
