### 2018-08-19 a17b9a2
enabling more generic use of the vanilla image reader without additional constraints
