### 2018-08-19 8493144
updating the requirements for clustering
