+++ requirements.txt
@@ -1,3 +1,4 @@
 nibabel
 matplotlib
 numpy
+scipy
