### 2018-08-19 413bb1f
basic tests for carpet class
