### 2018-08-19 20614a8
improved class presentation
