+++ docs/classes.rst
@@ -5,12 +5,45 @@ Visualization Classes
 A detailed usage illustration is available in the following `demo notebook <https://nbviewer.jupyter.org/github/raamana/mrivis/blob/master/docs/example_notebooks/mrivis_demo_Collage_SlicePicker_classes.ipynb>`_
 
 
-API Reference for Classes
-----------------------------
+SlicePicker API
+----------------
 
-.. automodule:: mrivis
-   :members: SlicePicker, MiddleSlicePicker, Collage, MidCollage, Carpet
+.. autoclass:: mrivis.SlicePicker
+   :members:
    :undoc-members:
    :show-inheritance:
 
 
+MiddleSlicePicker API
+---------------------
+
+.. autoclass:: mrivis.MiddleSlicePicker
+   :members:
+   :show-inheritance:
+
+
+Collage API
+----------------
+
+.. autoclass:: mrivis.Collage
+   :members:
+   :undoc-members:
+   :show-inheritance:
+
+
+MidCollage API
+----------------
+
+.. autoclass:: mrivis.MidCollage
+   :members:
+   :undoc-members:
+   :show-inheritance:
+
+
+Carpet API
+----------------
+
+.. autoclass:: mrivis.Carpet
+   :members:
+   :undoc-members:
+   :show-inheritance:
