+++ matlab_code/mrivis_collage.py
@@ -1,136 +1,0 @@
-# Autogenerated with SMOP 0.32-7-gcce8558
-from smop.core import *
-# mrivis_collage.m
-
-    
-@function
-def mrivis_collage(img_spec=None,fig_handle=None,rescale_intensity_range=None,varargin=None,*args,**kwargs):
-    varargin = mrivis_collage.varargin
-    nargin = mrivis_collage.nargin
-
-    # mrivis_collage( img, fig_handle)
-#   img  = MR image to be visualized
-#   fig_handle = figure handle to display the images
-#   scale_intensity_flag - whether to rescale image intensities or not
-#       can be a flag ([true]/false)
-#       or an array specifying the [ min max ] to which intensities should be rescaled to
-#       - This useful for comparison purposes:
-#           when images being visualized is exactly what you wanna see and not the rescaled ones
-#   varargin: optional text to be displayed
-    
-    # setting arguments
-    if nargin < 2:
-        fig_handle=copy(figure)
-# mrivis_collage.m:14
-    
-    if nargin < 3:
-        rescale_intensity_range=copy(true)
-# mrivis_collage.m:18
-    
-    # reading in data
-    if ischar(img_spec):
-        img_vol=MRIread(img_spec)
-# mrivis_collage.m:23
-        if numel(size(img_vol.vol)) != 3:
-            error('Input volume is not 3D!')
-        img=img_vol.vol
-# mrivis_collage.m:27
-    else:
-        if isreal(img_spec):
-            if numel(size(img_spec)) != 3:
-                error('Input volume is not 3D!')
-            img=copy(img_spec)
-# mrivis_collage.m:32
-        else:
-            display('Invalid input specified. Input either a path to image data, or provide 3d Matrix directly.')
-    
-    # cropping the image to its extents
-    padding=5
-# mrivis_collage.m:38
-    coords(arange(),1),coords[:,2],coords[:,3]=ind2sub(size(img),find(img > 0),nargout=3)
-# mrivis_collage.m:39
-    beg_coords=max(1,min(coords) - padding)
-# mrivis_collage.m:40
-    end_coords=min(size(img),max(coords) + padding)
-# mrivis_collage.m:41
-    img=img[beg_coords[1]:end_coords[1],beg_coords[2]:end_coords[2],beg_coords[3]:end_coords[3]]
-# mrivis_collage.m:42
-    num_cross_sections=11
-# mrivis_collage.m:44
-    slices=cellarray([[round(linspace(1,size(img,1),num_cross_sections))],[round(linspace(1,size(img,2),num_cross_sections))],[round(linspace(1,size(img,3),num_cross_sections))]])
-# mrivis_collage.m:45
-    # by default
-    RescaleImages=copy(true)
-# mrivis_collage.m:53
-    # estimating intensity ranges
-    if length(rescale_intensity_range) == 1 and rescale_intensity_range:
-        img_intensity_range=matlabarray(cat(min(ravel(img)),max(ravel(img))))
-# mrivis_collage.m:57
-    else:
-        if length(rescale_intensity_range) == 2:
-            img_intensity_range=copy(rescale_intensity_range)
-# mrivis_collage.m:59
-        else:
-            RescaleImages=copy(false)
-# mrivis_collage.m:61
-    
-    set(0,'CurrentFigure',fig_handle)
-    set(fig_handle,'Color','k')
-    for dim_index in arange(1,3).reshape(-1):
-        for range_index in arange(2,num_cross_sections - 1).reshape(-1):
-            # making the axis
-            subplot('Position',get_subplot_pos(dim_index,range_index))
[…]
