### 2018-08-22 d99e398
updating the additional requirements
