### 2018-08-26 #897 @miykael
Perfect! I have a conference and workshop coming up in the next 10 days but will start reviewing in the time between and finalize the review shortly after.
