### 2018-09-16 #897 @raamana
Yes it is. Sorry I should have made that explicit.
