### 2018-09-23 #897 @miykael
@arokem and @raamana I apologize for my late reply, I had an unforeseen hectic two weeks. But the review has now my full attention!

The software looks very good and useful and I will write my full review in the issue tracker of https://github.com/raamana/mrivis. But I think the repo is missing a `paper.md` file (as mentioned
in the [author guidlines](https://joss.readthedocs.io/en/latest/submitting.html#what-should-my-paper-contain), or is this optional?
