### 2018-09-23 #897 @raamana
thanks - looking fwd to learn your comments.

its in the docs folder, as required : https://github.com/raamana/mrivis/blob/master/docs/paper.md
