### 2018-09-23 raamana/mrivis#11 @miykael
Review for Journal of Open Source software
Here I will write my review for the JOSS submission: https://github.com/openjournals/joss-reviews/issues/897

# General impression
The purpose of the project is simple but very important and can be divided into two parts:
1. The comparison of two 3D MRI images. This is achieved with the **Alignment check** functions.
2. The creation of adjustable visualization collages for 3D MRI images, as well as the creation of carpet plots for 4D images. This is achieved with **Visualization Classes**.

**Alignment check** between MRI images is very important in neuroimaging. Reducing the similarity between two images to a simple number (e.g. using a distance function) might be accurate but is often times not useful as it doesn't provide information about the source of the discrepancy. Using a visual approach, and in this case, providing multiple ones is a very nice solution to this problem!

The creation of MRI **collages** is also very important for the field of neuroimaging. A collage allows the investigation of a 3D image (with many 100 of slices in three directions). All at once with a quick glance. Even though this is important, there are only a few open source software that provides this functionality, but none does it as simple and flexible as `mrivis`! The possibility of only plotting slices with a particular non-zero density or containing a given ROI is really great! A similar "one view check" can be achieved for 4D images, using the Carpet class!

My general impression of the repo is **very good**! The repo has an informative `README.rst` file, a very detailed documentation homepage, and provides important files, such as `LICENSE` `AUTHORS.rst`, `CODE_OF_CONDUCT.md`, `CONTRIBUTING.rst`, etc.

The code is well written, uses PEP440 standard (enforced with `flake8`) and is well commented.

The installation of the software is very easy. `mrivis` can be installed very quickly via `pip` and all required dependencies are listed in `requirements.txt`.

To quickly test `mrivis` and getting used to its functionality, this repo provides some example files as well as a demo notebook. This is very good, as it allows a user to quickly learn how to use `mrivis`.

# Review criteria

Concerning the review criteria under https://joss.readthedocs.io/en/latest/review_criteria.html:
- [x] Software license is **Acceptable**
- [x] Documentation is **very good**
- [x] A statement of need is **clearly defined**
- [x] Installation instructions are **Good**
- [x] Example usage are **included** (with example data)
- [x] API documentation is **Good**
- [x] Community guidelines are **provided**
- [x] Functionality is **Good** (minor bug issues mentioned below)
- [x] Tests are **Good** (done with Travis-CI)

# Additional Comments

I haven't reviewed the MATLAB code section in the [matlab_code](https://github.com/raamana/mrivis/tree/master/matlab_code) folder, as it contains the note: ***This is no longer maintained! The python version has more options and is recommended.***
