### 2018-09-23 raamana/mrivis#11 @raamana
Thanks Michael for taking the time to review this in great detail.

Does this mean you accept the software as is publication? Or are you planning to open any new issues for your suggestions?
