### 2018-09-23 raamana/mrivis#11 @miykael
Sorry @raamana, you were a bit quicker than I expected :-) I have a list of minor revisions and comments that I will post in the next hour but I wanted to separate it from my general feedback.
