### 2018-09-23 raamana/mrivis#11 @miykael
The following are my comments and issues that I encountered while reviewing `mrivis`. I apologize if it is too thorough and I want to iterate again that I think that the software is very well written, has a nice documentation and is clearly needed.

## Minor Revision

1. `paper.md` contains two authors but the list of authors in `AUTHORS.rst` and on zenodo only contains one author. This should be corrected.
1. `paper.md` contains minor typos: laboraries, sptial alignment, carefully-deisgned
1. `setup.cfg` contains `"current_version = 0.1.0"`, this should be updated to the current version 0.3.0. `HISTORY.rst` only contains version 0.1.0. For completeness, I recommend updating this to version 0.3.0 and `mrivis` is listed on PyPI ([here](https://pypi.org/project/mrivis/)) as version 0.3.5 but everywhere else as 0.3.0. Make sure that the versions are everywhere the same.
1. `requirements.txt` contains `scipy`, which is not mentioned under requirements in the `setup.py` file.
1. I highly recommend to also include a demo of the `Carpet` class in a notebook. This functionality is very useful! I additionally recommend to also include an example to the `cluster_rows_in_roi` method.
1. Should the carpet plot approach be inspired or related to the paper from [Power et al. (2016)](https://www.sciencedirect.com/science/article/pii/S1053811916303871) then I recommend to cite it in `paper.md` as well.
1. The `color_mix` function doesn't seem to work correctly. Using the following code:
    ```python
    path1 = '../../example_datasets/3569_bl_PPMI.nii'
    from mrivis import color_mix
    color_mix(path1, path1);
    ```
    I get the following error: `TypeError: '<' not supported between instances of 'tuple' and 'int'`
1. In the **Alignment check - usage** section, using the following code in a terminal doesn't show any figure/image:
    ```python
    from mrivis import checkerboard

    path1 = '/Users/Reddy/Desktop/image.nii'
    path2 = '/Users/Reddy/Desktop/another.nii'

    checkerboard(path1, path2) # square patches
    ```
    There seems to be something missing. Is it assumed that people have already imported `matplotlib` in a particular way? Do we need to use the `.show()` method afterward?
1. `mrivis` contains additional `utils` functions that are not included on the documentation homepage, such as `scale_0to1`, `threshold_image` and `row_wise_rescale`. I recommend to add them to the API documentation as well.
1. Running the example in the **Alignment check - usage** section doesn't necessarily lead to nice looking figures as one might need to adapt additional parameters, such as `background_threshold` or `view_set`. I recommend having a working demo notebook that shows the functionality with on the provided example data.
1. Some of the methods in the **Visualization class** have input parameters that are not listed in the docstring (see [here](https://raamana.github.io/mrivis/classes.html#mrivis.SlicePicker.get_slices) or [here](https://raamana.github.io/mrivis/classes.html#mrivis.Collage.attach)).
1. [`Carpet` class](https://raamana.github.io/mrivis/classes.html#carpet-api) contains the input parameter `num_frames_to_skip` that isn't described in the docstring.
1. The `color_mix` function from `workflow.py` has parameter description triplets (i.e. `cmap`, `overlay_image` and `overlay_alpha` are explained three times). This even though none of these parameters are actual input parameters to this function. It seems that the description text of those three parameters belong to the `voxelwise_diff` function.
1. The `color_mix` function is missing the description text in the docstring for the parameter `color_space`.
1. `voxelwise_diff` function accepts an input parameter `overlay_image` that is not described in the doc string.
1. Functions in `color_maps.py` are never used in any example/demo, isn't tested or mentioned in the documentation. It seems that they are only needed for the `aseg_on_mri` function, defined in `workflow.py` which is also neither tested, documented or explained.
1. The following functions are defined but never used in the code base: `check_bounding_rect`, `get_freesurfer_subcortical_LUT`, `_open_figure`
1. `README.rst` on github contains the section ***If you would like to use the export to GIF from SlicePicker, then install the imageio package as well via pip install imageio***. This section is not included in the docs under **installation**. I recommend to add this section to the documentation. Perhaps also include an example on how to create a GIF in the demo notebook.

## Tutorial/Documentation revision

1. The current version of [`mrivis_demo_Collage_SlicePicker_classes.ipynb`](https://nbviewer.jupyter.org/github/raamana/mrivis/blob/master/docs/example_notebooks/mrivis_demo_Collage_SlicePicker_classes.ipynb) contains the error `No module named 'mrivis'`. I suggest correcting this by just rerunning the notebook.
1. `mrivis_demo_Collage_SlicePicker_classes.ipynb` is missing the `import numpy as np` command.
1. For the plotting of two subplots in `mrivis_demo_Collage_SlicePicker_classes.ipynb` to work, the two cells...
    ```python
    # lets have two subplots
    fig, ax = plt.subplots(2, 1, squeeze=True)
    ```

    and

    ```python
    # you just need to pass in all the
    for s1, s2 in sp.get_slices_multi([img_one, img_two]):
        ax[0].imshow(s1)
        ax[1].imshow(s2)
        sleep(0.1)
        plt.pause(0.001)
    #
    # PS: You might not be able to see the animation in the jupyter notebook.
    ```

    need to be combined. Otherwise, this doesn't work in a notebook.
1. The collage command in `mrivis_demo_Collage_SlicePicker_classes.ipynb` does not work in a notebook if the three first cells are separated. The first cell creates an empty canvas that cannot be accessed by the following cells. Either combine the three cells into one or replace `plt.show(collage.fig)` with ` collage.fig`.
1. The notebook contains the phrase: *"Coming: vary the number of slices per each view (this is implemented, needs to be tested more)"*. I recommend to remove this sentence and to open a corresponding issue on github.
1. The last command (`bounding_rect`) in `mrivis_demo_Collage_SlicePicker_classes.ipynb` doesn't seem to work correctly in the notebook. The collage is not constraint to the top right corner. The command seems to work if the Python code is run from a terminal, but the collage doesn't seem to be sized correctly neither, as the right part of the images is cut off/reaches over the canvas.

## Optional changes

1. Add an example figure for the collage, as well as for the `Carpet` function for 4D images to `README.rst` to show the full functionality of the software.
1. The demo notebook(s) could also be made available via [binder.org](https://mybinder.org/). Such a service would allow showcasing `mrivis`'s functionality directly on the web. No additional setup would be required.
1. The documentation homepage has a **usage** and **API** section for the **alignment check** but only an **API** section for the **Visualization classes**. I would recommend having also a **usage** section for the **Visualization classes**
1. I recommend to also create an example/demo notebook for the alignment check functionality of `mrivis`.
1. The codebase contains 14 `TODO` comments. I recommend to either solve this todo's or create individual issues for them. This helps to quickly finish those points or to make others aware of them.
