### 2018-09-23 #897 @miykael
@raamana and @arokem - my review can be found under https://github.com/raamana/mrivis/issues/11. @raamana, don't hesitate to ask for clarification!
