### 2018-09-24 58c56b9
dealing with future warning - numpy slicing
