### 2018-09-24 5fd58a0
improving API docs
