### 2018-09-25 raamana/mrivis#11 @raamana
Thanks Michael, this is pretty detailed and helpful. I'm revising it now and should be done soon.

Few questions:
 - option of bounding_rect in Collage is buggy - I spent sometime before trying to make it more reliable, but found matplotlib docs to be not helpful enough and I anticipate this may MUCH MORE effort to debug. Would it okay if I list it as a known issue for  now (in docs/API etc) and flag it for future debugging.
 - `color_maps` module is used by other libraries such as `visualqc`, and I included them here as they are more visualization oriented than QC oriented. Perhaps mentioning them in the list of features would be helpful?
