### 2018-09-25 raamana/mrivis#11 @miykael
Happy to hear that it helps!

- I think the `bounding_rect` (almost?) worked if it's run directly from Python code (i.e. not with Jupyter notebook). Therefore, if you flag it as a future debug issue and just mention in the demo notebook that it doesn't work in the notebook, that's alright with me. It's a really cool feature but not essential for the toolbox to be shipped.
- Sorry, I didn't consider that `color_maps` might be needed by other libraries. I have no problem if you keep it in there, I just wanted to point out that this might be either some legacy code or hasn't grown to its full potential yet. So if you mention this as an additional feature, even better :+1:
