### 2018-10-01 acb71e5
cleaning up setup files
