### 2018-10-01 d186210
updating the paper with citations and typo fixes
