+++ docs/paper.bib
@@ -25,10 +25,11 @@
 }
 @misc{raamana2018visualqc,
   author       = {Pradeep Reddy Raamana},
-  title        = {{VisualQC: Assistive tools for easy and rigorous 
+  title        = {{VisualQC: Assistive tools for easy and rigorous
                    quality control of neuroimaging data}},
   month        = apr,
   year         = 2018,
   doi          = {10.5281/zenodo.1211365},
   url          = {https://doi.org/10.5281/zenodo.1211365}
 }
+@article{power2017carpet, title={A simple but useful way to assess fMRI scan qualities}, volume={154}, ISSN={10538119}, DOI={10.1016/j.neuroimage.2016.08.009}, journal={NeuroImage}, author={Power, Jonathan D.}, year={2017}, month={Jul}, pages={150–158} }
