### 2018-10-01 aab1305
much improved docs
