+++ docs/API.rst
@@ -8,3 +8,13 @@ A tutorial-like presentation is available at :doc:`usage`, using the following A
    :members: checkerboard, voxelwise_diff, color_mix
    :undoc-members:
    :show-inheritance:
+
+
+Helpers and Utility scripts
+-----------------------------
+
+.. automodule:: mrivis.utils
+   :members: scale_0to1, row_wise_rescale, threshold_image
+   :undoc-members:
+   :show-inheritance:
+
