### 2018-10-02 aed0e00
bugfix: carpet clustering with defaults, updated tests
