### 2018-10-02 8c58ac7
re-adding slicepicker to gif in master (prev PR merge disappeared)
