### 2018-10-02 3dc51ed
renamed notebook, separating API from usage
