+++ docs/vis_classes_api.rst
@@ -1,9 +1,6 @@
 ---------------------------
-Visualization Classes
+Visualization Classes - API
 ---------------------------
-
-A detailed usage illustration is available in the following `demo notebook <https://nbviewer.jupyter.org/github/raamana/mrivis/blob/master/docs/example_notebooks/mrivis_demo_Collage_SlicePicker_classes.ipynb>`_
-
 
 SlicePicker API
 ----------------
