+++ docs/usage.rst
@@ -4,4 +4,4 @@ Usage
 
 For usage instructions on how to use `mrivis`
  - to check alignment, see :doc:`alignment_usage`
- - to use different classes, see :doc:`classes`
+ - to use different classes, see :doc:`vis_classes_api`
