### 2018-10-02 eb681f9
improved docs
