+++ docs/usage.rst
@@ -3,5 +3,6 @@ Usage
 ------
 
 For usage instructions on how to use `mrivis`
- - to check alignment, see :doc:`alignment_usage`
- - to use different classes, see :doc:`vis_classes_api`
+ - to check alignment, see :doc:`alignment_usage` and :doc:`alignment_api`
+ - to use different classes, see :doc:`vis_classes_usage` and :doc:`vis_classes_api`
+
