+++ README.rst
@@ -24,6 +24,11 @@ A sample gallery of different ways to check alignment is shown below:
 Full documentation https://raamana.github.io/mrivis/readme.html
 
 
+Executable and interactive demo:
+
+.. image:: https://mybinder.org/badge.svg :target: https://mybinder.org/v2/gh/raamana/mrivis/master?filepath=docs%2Fexample_notebooks%2Fmrivis_demo_vis_classes.ipynb
+
+
 **Installation**
 
 .. code-block:: console
