+++ docs/vis_classes_usage.rst
@@ -1,0 +1,9 @@
+-----------------------------
+Visualization Classes - Usage
+-----------------------------
+
+A detailed usage illustration is available in the following `demo notebook <https://nbviewer.jupyter.org/github/raamana/mrivis/blob/master/docs/example_notebooks/mrivis_demo_vis_classes.ipynb>`_
+
+Executable and interactive demo:
+
+.. image:: https://mybinder.org/badge.svg :target: https://mybinder.org/v2/gh/raamana/mrivis/master?filepath=docs%2Fexample_notebooks%2Fmrivis_demo_vis_classes.ipynb
