+++ _sources/index.rst.txt
@@ -11,7 +11,8 @@ Contents:
    usage
    alignment_usage
    alignment_API
-   classes
+   vis_classes_usage
+   vis_classes_api
    contributing
    authors
 
