+++ _sources/usage.rst.txt
@@ -3,5 +3,6 @@ Usage
 ------
 
 For usage instructions on how to use `mrivis`
- - to check alignment, see :doc:`alignment_usage`
- - to use different classes, see :doc:`classes`
+ - to check alignment, see :doc:`alignment_usage` and :doc:`alignment_API`
+ - to use different classes, see :doc:`vis_classes_usage` and :doc:`vis_classes_api`
+
