+++ _sources/vis_classes_api.rst.txt
@@ -1,0 +1,46 @@
+---------------------------
+Visualization Classes - API
+---------------------------
+
+SlicePicker API
+----------------
+
+.. autoclass:: mrivis.SlicePicker
+   :members:
+   :undoc-members:
+   :show-inheritance:
+
+
+MiddleSlicePicker API
+---------------------
+
+.. autoclass:: mrivis.MiddleSlicePicker
+   :members:
+   :show-inheritance:
+
+
+Collage API
+----------------
+
+.. autoclass:: mrivis.Collage
+   :members:
+   :undoc-members:
+   :show-inheritance:
+
+
+MidCollage API
+----------------
+
+.. autoclass:: mrivis.MidCollage
+   :members:
+   :undoc-members:
+   :show-inheritance:
+
+
+Carpet API
+----------------
+
+.. autoclass:: mrivis.Carpet
+   :members:
+   :undoc-members:
+   :show-inheritance:
