### 2018-10-02 c6cb5e7
improved docs
