+++ docs/vis_classes_usage.rst
@@ -6,4 +6,4 @@ A detailed usage illustration is available in the following `demo notebook <http
 
 Executable and interactive demo:
 
-.. image:: https://mybinder.org/badge.svg :target: https://mybinder.org/v2/gh/raamana/mrivis/master?filepath=docs%2Fexample_notebooks%2Fmrivis_demo_vis_classes.ipynb
+.. image:: https://mybinder.org/badge.svg :target: https://hub.mybinder.org/user/raamana-mrivis-9y51wqyz/notebooks/docs/example_notebooks/mrivis_demo_vis_classes.ipynb
