### 2018-10-02 dbc5096
improved docs
