### 2018-10-02 d7e3ca0
improved docs
