+++ README.rst
@@ -26,7 +26,8 @@ Full documentation https://raamana.github.io/mrivis/readme.html
 
 Executable and interactive demo:
 
-.. image:: https://mybinder.org/badge.svg :target: https://mybinder.org/v2/gh/raamana/mrivis/master?filepath=docs%2Fexample_notebooks%2Fmrivis_demo_vis_classes.ipynb
+.. image:: https://mybinder.org/badge.svg
+    :target: https://mybinder.org/v2/gh/raamana/mrivis/master?filepath=docs%2Fexample_notebooks%2Fmrivis_demo_vis_classes.ipynb
 
 
 **Installation**
