### 2018-10-02 16a3401
add more funcs to API
