+++ docs/alignment_API.rst
@@ -9,3 +9,12 @@ A tutorial-like presentation is available at :doc:`alignment_usage`, using the f
    :members: checkerboard, voxelwise_diff, color_mix
    :undoc-members:
    :show-inheritance:
+
+Helpers and Utility scripts
+-----------------------------
+
+.. automodule:: mrivis.utils
+   :members: scale_0to1, row_wise_rescale, threshold_image
+   :undoc-members:
+   :show-inheritance:
+
