### 2018-10-03 raamana/mrivis#11 @miykael
hey @raamana - great! I think we can finish this rather quickly.

## Needs to be addressed

1. The `color_mix` function (as described [here](https://raamana.github.io/mrivis/alignment_usage.html?highlight=color_mix#alignment-check-usage) still doesn't seem to work correctly. Should the following code work or do I need to run it differently?
   ```python
    path1 = 'example_datasets/3569_bl_PPMI.nii'
    path2 = 'example_datasets/4086_bl_PPMI.nii'
    from mrivis import color_mix
    color_mix(path1, path2);
    ```
1. `mrivis_demo_vis_classes.ipynb` is still missing the `numpy as np` import. Without it, it will not finish as it is.
1. What's the difference between `mrivis_demo_Collage_SlicePicker_classes.ipynb` and `mrivis_demo_vis_classes.ipynb`? The notebooks seems to be exactly the same. If they are the same, you should delete one and make sure to only link to this one throughout the documentation.

## Comments about the revisions

1. The new carpet plot notebook example is very nice!
1. Having an alignment notebook would be very nice. But you describe very well how one can use it in the documentation, so it's not a requirement for the paper.
1. Stephen C. Strother is still not mentioned under `AUTHOR.rst` or on [zenodo](https://zenodo.org/record/1328279). But as you mention him in the JOSS paper, that's ok for me.
1. The `color_mix` function still contains the docstrings for the parameters `cmap`, `overlay_image`, `overlay_alpha` which are no valid input parameters. You should delete those descriptions from the docstring.
1. About the unused methods, all fine with me.
1. It's ok for me if you leave the TODO in the code, I understand this approach very well :-) Just thought that if you create issues, others might see the problematic and might be able to work on it themselves. Also, I wanted that you verify that those are not relevant for the JOSS paper.

## Concerning mybinder (offtopic)

**Off-topic:**
What do you mean about `nbviwer` is not reliable? I never had problems with it. But yes, with `mybinder`, all the time.

As an additional note, your current notebook will not work in mybinder because you include your personal absolute path `base_dir = '/Users/Reddy/dev/mrivis/example_datasets'`. I recommend to change this in the notebook to this `base_dir = '../../example_datasets/'`. Like that, the mybinder notebook will also work without applying any changes.
