### 2018-10-03 80984da
more instructions on changing the paths
