### 2018-10-03 raamana/mrivis#11 @raamana
Thanks Michael for the prompt and thorough checks.. Sorry, I thought I made the changes necessary for the above, apparently not.

## Needs to be addressed

1. The `color_mix` function (as described [here](https://raamana.github.io/mrivis/alignment_usage.html?highlight=color_mix#alignment-check-usage) is not supposed to work as shown. I now left more instructions on how to make it work

1. Added the `numpy` import to notebook now

1. there is NO difference between `mrivis_demo_Collage_SlicePicker_classes.ipynb` and `mrivis_demo_vis_classes.ipynb`, just duplicates created from bad merge

## Comments about the revisions

1. Thanks - carpet plot notebook needs to be significantly expanded in the future with real world use-cases, and with additional functionality such as blurring within ROI etc. Some other options I plan to add are composite layering of stats (as in visualqc), and mixing multiple variations of carpet (temporal derivatives etc).This would require significant efforts, and I am looking for more time (on my part) and potential contributors. Just opened issues for these ideas.
1. Okay
1. Stephen C. Strother is still not mentioned under `AUTHORS.rst` as its more a list of who contributed code (and as it is mentioning only Dev Lead). SCS has contributed to design, conception of useful functionality and financially supporting the work. He has not contributed code yet, so didn't want to add his name. I could perhaps delete this file if its confusing.
1. removed the invalid params from docstrings
1. great
1. Just checked TODOs - they do not interfere with the advertised functionality. They are more of a future TODOs for me.

**Off-topic:**

- nbviewer was throwing the `404 not found` error - likely due to delays in syncing across GitHub and nbviewer
- fixed the hardcoded paths in notebook, it should now work on mybinder (after they get synced and relaunched)
