### 2018-10-03 raamana/mrivis#11 @miykael
Thank you @raamana - this seems all good to me!

Irrelevant for the paper, but just that you know. The deleted `mrivis_demo_Collage_SlicePicker_classes.ipynb` string is still appearing under
- `docs/example_notebooks/mrivis_demo_vis_classes.ipynb`
- `docs/readme.rst`
- `docs/classes.rst`
