### 2018-10-03 raamana/mrivis#11 @raamana
do you see the deleted notebook here: https://github.com/raamana/mrivis/tree/master/docs/example_notebooks ?
