### 2018-10-03 raamana/mrivis#11 @miykael
No, the deleted notebook is gone. But when I do a `grep -r "mrivis_demo_Collage_SlicePicker_classes"` in the downloaded source code, I'll still find the old notebook at those three places.
