### 2018-10-03 raamana/mrivis#11 @raamana
They are hidden files and checkpoints (not a concern for users).. I've updated `.rst`'s on gh-pages branch (which is what goes on the docs website), but I will try clean them up on master too.

```
$ 10:28:15 SQuark-3 mrivis >>  grep -rl "mrivis_demo_Collage_SlicePicker_classes" .
./docs/example_notebooks/mrivis_demo_vis_classes.ipynb
./docs/example_notebooks/.ipynb_checkpoints/mrivis_demo_Collage_SlicePicker_classes-checkpoint.ipynb
./docs/example_notebooks/.ipynb_checkpoints/mrivis_demo_vis_classes-checkpoint.ipynb
./docs/classes.rst
./docs/_build/html/_sources/readme.rst.txt
./docs/_build/html/readme.html
./docs/_build/doctrees/readme.doctree
./docs/readme.rst
$ 10:28:25 SQuark-3 mrivis >>
```
