### 2018-10-03 raamana/mrivis#11 @miykael
Sorry, I didn't mean the files in `.ipynb_checkpoints` or `_build`  but for example `docs/readme.rst` on master that creates a nbviewer 404 error [here](https://github.com/raamana/mrivis/blob/master/docs/readme.rst) if you click on the demo button.
I was just worried that they have a direct influence on gh-pages.
