+++ setup.cfg
@@ -8,7 +8,7 @@ tag_prefix =
 parentdir_prefix = mrivis-
 
 [bumpversion]
-current_version = 0.3.5
+current_version = 0.3.6
 commit = True
 tag = True
 
