### 2018-10-03 raamana/mrivis#11 @raamana
Thanks for spotting it - fixed the link in `docs/readme.rst` in master now, although I hope users will go through the website only :)
