### 2018-10-03 #897 @raamana
Just pushed the latest code to pypi also! :)
