### 2018-10-05 #897 @arokem
@raamana : sorry - a couple of things I noticed about your article just now:

- The title is not capitalized. I don't think that you want every word in the title to be capitalized, but the very first word probably should be. Also, the title currently doesn't include the name of the software library. I would recommend adding that.

- The references do not have DOIs. Please add these.
