### 2018-10-05 02789d6
updated paper with DOIs etc
