+++ docs/paper.bib
@@ -5,7 +5,8 @@
   volume={9},
   number={3},
   year={2007},
-  publisher={IEEE}
+  publisher={IEEE},
+  doi={10.1109/MCSE.2007.58}
 }
 @article{hunter2007matplotlib,
   title={Matplotlib: A 2D graphics environment},
@@ -15,13 +16,17 @@
   number={3},
   pages={90--95},
   year={2007},
-  publisher={IEEE}
+  publisher={IEEE},
+  doi={10.1109/MCSE.2007.55},
 }
+
 @article{brett2016nibabel,
   title={nibabel: Access a cacophony of neuro-imaging file formats, version 2.1. 0},
   author={Brett, Matthew and Hanke, Michael and Cipollini, Ben and C{\^o}t{\'e}, Marc-Alexandre and Markiewicz, Chris and Gerhard, Stephan and Larson, Eric and Lee, Gregory R and Halchenko, Yaroslav and Kastman, Erik and others},
   journal={Zenodo},
-  year={2016}
+  year={2016},
+  doi={10.5281/zenodo.591597}
+
 }
 @misc{raamana2018visualqc,
   author       = {Pradeep Reddy Raamana},
@@ -32,4 +37,12 @@
   doi          = {10.5281/zenodo.1211365},
   url          = {https://doi.org/10.5281/zenodo.1211365}
 }
-@article{power2017carpet, title={A simple but useful way to assess fMRI scan qualities}, volume={154}, ISSN={10538119}, DOI={10.1016/j.neuroimage.2016.08.009}, journal={NeuroImage}, author={Power, Jonathan D.}, year={2017}, month={Jul}, pages={150–158} }
+@article{power2017carpet,
+title={A simple but useful way to assess fMRI scan qualities},
+volume={154}, ISSN={10538119},
+DOI={10.1016/j.neuroimage.2016.08.009},
+journal={NeuroImage},
+author={Power, Jonathan D.},
+year={2017},
+month={Jul},
+pages={150–158} }
